Removal of internal units.
* Remove internal units from BASE_SCREEN and it's derivatives.
* Remove internal units from EDA_DRAW_FRAME and it's derivatives.
* Use build time code to replace internal units conversions.
* Fix scaling bug in page layout sample window that I created in my
  last commit.

+++ include/base_units.h
@@ -96,7 +96,7 @@ void PutValueInLocalUnits( wxTextCtrl& aTextCtr, int aValue );
 /**
  * Return in internal units the value "val" given in inch or mm
  */
-int From_User_Unit( EDA_UNITS_T aUnit, double aValue );
+double From_User_Unit( EDA_UNITS_T aUnit, double aValue );
 
 /**
  * Function ReturnValueFromeString
